[SKP-73] refac : 분석, 상세 화면 컴포넌트화 및 스타일 분리
[SKP-73] s
