+++ src/navigators/StackNavigator.tsx
@@ -3,9 +3,9 @@ import TabNavigator from '../navigators/TabNavigator';
 import {StackMenu} from '../navigators/constants/menu';
 import {StackParamList} from '../navigators/types';
 import Home from '../screens/Home';
-import Analyze from '../screens/Analyze';
 import AnalyzeResult from '../screens/AnalyzeResult/AnalyzeResult';
-import Detail from '../screens/Detail';
+import Detail from '../screens/Detail/Detail';
+import Analyze from '../screens/Analyze/Analyze';
 
 const Stack = createNativeStackNavigator<StackParamList>();
