+++ src/screens/AnalyzeResult/dummies.json
@@ -1,31 +1,33 @@
 {
   "results": [
-    {
-      "title": "인천대공원",
-      "address": "인천 남동구 장수동 산 79",
-      "category": "자연"
-    },
-    {
-      "title": "을왕리 해수욕장",
-      "address": "인천광역시 서동구 비비로 123",
-      "category": "익사이팅"
-    },
-    {
-      "title": "인천대공원",
-      "address": "인천 남동구 장수동 산 79",
-      "category": "자연"
-    },
-    {
-      "title": "을왕리 해수욕장",
-      "address": "인천광역시 서동구 비비로 123",
-      "category": "익사이팅"
-    }
-  ],
-  "result": [
-    {
-      "title": "인천대공원",
-      "address": "인천 남동구 장수동 산 79",
-      "category": "자연"
-    }
+    [
+      {
+        "title": "인천대공원",
+        "address": "인천 남동구 장수동 산 79",
+        "category": "자연"
+      },
+      {
+        "title": "을왕리 해수욕장",
+        "address": "인천광역시 서동구 비비로 123",
+        "category": "익사이팅"
+      },
+      {
+        "title": "인천대공원",
+        "address": "인천 남동구 장수동 산 79",
+        "category": "자연"
+      },
+      {
+        "title": "을왕리 해수욕장",
+        "address": "인천광역시 서동구 비비로 123",
+        "category": "익사이팅"
+      }
+    ],
+    [
+      {
+        "title": "인천대공원",
+        "address": "인천 남동구 장수동 산 79",
+        "category": "자연"
+      }
+    ]
   ]
 }
